Tests/LibWeb/Text/expected/HTML/FormData-append-blob-with-type.txt

Name: test.txt
Type: text/plain
